refactor: Consolidate agent state unit tests
